# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  username: string;
  isAdmin: boolean;
}

export interface UserInfo {
  username: string;
  isAdmin: boolean;
  isLoggedIn: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface HeroContent {
  title: string;
  subtitle: string;
  description: string;
  profileImageUrl?: string;
  titleFontSize?: string;
  subtitleFontSize?: string;
  descriptionFontSize?: string;
}

export interface ProjectSubSection {
  title: string;
  items: string[];
}

export interface ProjectSection {
  heading: string;
  items: string[];
  subSections?: ProjectSubSection[];
}

export interface ProjectImage {
  url: string;
  caption: string;
  position?: string;
}

export interface Project {
  id: number;
  title: string;
  summary: string;
  category: string;
  sortOrder: number;
  details: ProjectSection[];
  images: ProjectImage[];
  tags: string[];
  titleFontSize?: string;
  summaryFontSize?: string;
}

export interface ProjectInput {
  title: string;
  summary: string;
  category: string;
  sortOrder: number;
  details: ProjectSection[];
  images: ProjectImage[];
  tags: string[];
  titleFontSize?: string;
  summaryFontSize?: string;
}

export interface SkillsContent {
  sectionTitle: string;
  skills: string[];
  titleFontSize?: string;
}

export interface ClosingContent {
  text: string;
  fontSize?: string;
}

export interface PortfolioSettings {
  siteTitle: string;
  ownerName: string;
  baseFontSize?: string;
  accentColor?: string;
  showCareerSection?: boolean;
  showSkillsSection?: boolean;
  showClosingSection?: boolean;
}

export interface CareerEntry {
  period: string;
  company: string;
  role: string;
  description: string;
  sortOrder: number;
}

export interface CareerContent {
  sectionTitle: string;
  entries: CareerEntry[];
  titleFontSize?: string;
}