# lib/api-zod/src/generated/types/careerContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { CareerEntry } from "./careerEntry";

export interface CareerContent {
  sectionTitle: string;
  entries: CareerEntry[];
  titleFontSize?: string;
}