# lib/api-zod/src/generated/types/careerEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerEntry {
  period: string;
  company: string;
  role: string;
  description: string;
  sortOrder: number;
}