# lib/api-zod/src/generated/types/closingContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ClosingContent {
  text: string;
  fontSize?: string;
}