# lib/api-zod/src/generated/types/heroContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface HeroContent {
  title: string;
  subtitle: string;
  description: string;
  profileImageUrl?: string;
  titleFontSize?: string;
  subtitleFontSize?: string;
  descriptionFontSize?: string;
}