# lib/api-zod/src/generated/types/portfolioSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioSettings {
  siteTitle: string;
  ownerName: string;
  baseFontSize?: string;
  accentColor?: string;
  showCareerSection?: boolean;
  showSkillsSection?: boolean;
  showClosingSection?: boolean;
}