# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectImage } from "./projectImage";
import type { ProjectSection } from "./projectSection";

export interface Project {
  id: number;
  title: string;
  summary: string;
  category: string;
  sortOrder: number;
  details: ProjectSection[];
  images: ProjectImage[];
  tags: string[];
  titleFontSize?: string;
  summaryFontSize?: string;
}