# lib/api-zod/src/generated/types/projectImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectImage {
  url: string;
  caption: string;
  position?: string;
}