# lib/api-zod/src/generated/types/projectSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Science Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSubSection } from "./projectSubSection";

export interface ProjectSection {
  heading: string;
  items: string[];
  subSections?: ProjectSubSection[];
}